% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{add_observation_errors}
\alias{add_observation_errors}
\title{Create error_indicators using error_params}
\usage{
add_observation_errors(true_data, error_params, range_dates)
}
\description{
Create error_indicators using error_params
}
