% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_aug_data.R
\name{calculate_delay_boundaries}
\alias{calculate_delay_boundaries}
\title{Calculate the lower and upper delay boundaries based on quantiles}
\usage{
calculate_delay_boundaries(delay_params, quantile_range)
}
\arguments{
\item{delay_params}{The data frame of delay distribution parameters}

\item{quantile_range}{A vector of two probabilities (e.g. c(0.01, 0.99))}
}
\value{
A data frame with \code{from}, \code{to}, \code{group}, \code{min_delay}, and \code{max_delay}
}
\description{
Calculate the lower and upper delay boundaries based on quantiles
}
\examples{
delay_params <- data.frame(
  group = c(1:4, 2, 3, 3, 4, 4),
  from = c("onset", "onset", "onset", "onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "report", "report", "report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  delay_mean = c(10, 10, 10, 10, 15, 7, 20, 7, 12),
  delay_cv = c(0.3, 0.3, 0.3, 0.3, 0.4, 0.2, 0.5, 0.2, 0.3)
)
quantile_range <- c(0.01, 0.99)
calculate_delay_boundaries(delay_params, quantile_range)

}
