% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-initialise.R
\name{calculate_transitive_steps}
\alias{calculate_transitive_steps}
\title{Function to work out number of steps between events}
\usage{
calculate_transitive_steps(delay_map)
}
\arguments{
\item{delay_map}{Data frame defining all delay rules}
}
\value{
Data frame of the steps between each event date
}
\description{
Function to work out number of steps between events
}
\examples{
delay_map <- data.frame(
  from = c("onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  group = I(list(1:4, 2, 3, 3, 4, 4))
)

calculate_transitive_steps(delay_map)

}
