% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronofix_plot_delays.R
\name{chronofix_plot_delays}
\alias{chronofix_plot_delays}
\title{Plot Estimated Delay Distributions}
\usage{
chronofix_plot_delays(
  mcmc_output,
  delay_map,
  n_points = 200,
  facet_by_group = TRUE,
  share_x_axis = TRUE
)
}
\arguments{
\item{mcmc_output}{Output list from \code{chronofix_mcmc_run()}}

\item{delay_map}{The delay_map used for the model setup}

\item{n_points}{Number of points along the x-axis to evaluate (default 200)}

\item{facet_by_group}{Logical, if TRUE, creates a grid with a row per group
(default TRUE)}

\item{share_x_axis}{Logical, if TRUE, plots all distributions on the same
x-axis scale (default TRUE)}
}
\description{
Plot Estimated Delay Distributions
}
