% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{datefixer_hyperparameters}
\alias{datefixer_hyperparameters}
\title{Create hyperparameters}
\usage{
datefixer_hyperparameters(
  prob_error_shape1 = 1,
  prob_error_shape2 = 1,
  mean_delay_scale = 10,
  cv_delay_scale = 10
)
}
\arguments{
\item{prob_error_shape1}{The first shape parameter of the beta prior
distribution for the probability of error}

\item{prob_error_shape2}{The second shape parameter of the beta prior
distribution for the probability of error}

\item{mean_delay_scale}{The scale parameter (mean) of the exponential prior
distribution for the means of the delays}

\item{cv_delay_scale}{The scale parameter (mean) of the exponential prior
distribution for the coefficients of variations of the delays}
}
\value{
List of hyperparameters
}
\description{
Create hyperparameters
}
