% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{datefixer_model}
\alias{datefixer_model}
\title{Create a datefixer model}
\usage{
datefixer_model(data, delay_map, hyperparameters, control)
}
\arguments{
\item{data}{Observed data}

\item{delay_map}{Delays information}

\item{hyperparameters}{List of hyperparameters}

\item{control}{List of control parameters}
}
\value{
A datefixer model
}
\description{
Create a datefixer model
}
