% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{datefixer_sampler}
\alias{datefixer_sampler}
\title{Datefixer sampler}
\usage{
datefixer_sampler(control)
}
\arguments{
\item{control}{Object produced by \code{mcmc_control()}}
}
\value{
A monty sampler configured for the datefixer model.
}
\description{
Datefixer sampler
}
