% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_aug_data.R
\name{initialise_augmented_data}
\alias{initialise_augmented_data}
\title{Initialises augmented data based on observed data}
\usage{
initialise_augmented_data(
  observed_data,
  delay_map,
  delay_params,
  init_settings
)
}
\arguments{
\item{observed_data}{A data frame of observed dates}

\item{delay_map}{A data frame defining the delay rules}

\item{delay_params}{A data frame with delay distribution parameters}

\item{init_settings}{A list containing \code{quantile_range} (e.g. c(0.01, 0.99))}
}
\value{
A list with two data frames: \code{augmented_data} and \code{error_indicators}
}
\description{
Initialises augmented data based on observed data
}
\examples{
# Define the delay_map data frame
delay_map <- data.frame(
  from = c("onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  group = I(list(1:4, 2, 3, 3, 4, 4))
)

# Define the delay parameters data frame
delay_params <- data.frame(
  group = c(1:4, 2, 3, 3, 4, 4),
  from = c("onset", "onset", "onset", "onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "report", "report", "report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  delay_mean = c(10, 10, 10, 10, 15, 7, 20, 7, 12),
  delay_cv = c(0.3, 0.3, 0.3, 0.3, 0.4, 0.2, 0.5, 0.2, 0.3)
)

# Define other parameters
n_per_group <- rep(10, max(delay_params$group))
error_params <- list(prop_missing_data = 0.2, prob_error = 0.05)
range_dates <- as.integer(as.Date(c("2025-03-01", "2025-09-01")))

# Simulate data
set.seed(1)
sim_result <- simulate_data(
  n_per_group = n_per_group,
  delay_map = delay_map,
  delay_params = delay_params,
  error_params = error_params,
  range_dates = range_dates,
  simul_error = TRUE
)

obs_dat <- sim_result$observed_data

mcmc_init_settings <- list(quantile_range = c(0.01, 0.99))

aug_dat <- initialise_augmented_data(
  observed_data = obs_dat,
  delay_map = delay_map,
  delay_params = delay_params,
  init_settings = mcmc_init_settings
)

# Compare data to original
obs_dat
aug_dat$augmented_data
aug_dat$error_indicators

}
