% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-initialise.R
\name{initialise_row}
\alias{initialise_row}
\title{Initialise the augmented data for an individual (row)}
\usage{
initialise_row(individual_data, delay_map, delay_boundaries)
}
\arguments{
\item{individual_data}{A single row data frame for an individual}

\item{delay_map}{Data frame defining all delay rules}

\item{delay_boundaries}{Data frame with calculated min/max delays}
}
\value{
A single row of the data frame with dates corrected and imputed
}
\description{
Initialise the augmented data for an individual (row)
}
\examples{
# Inputs
individual_data <- data.frame(
  id = 1, group = 3,
  onset = as.Date("2025-05-01"),
  hospitalisation = as.Date("2025-05-02"),
  discharge = as.Date(NA)
)

delay_map <- data.frame(
  from = c("onset", "hospitalisation"),
  to = c("hospitalisation", "discharge"),
  group = I(list(3, 3))
)

delay_boundaries <- data.frame(
  group = 3,
  from = c("onset", "hospitalisation"),
  to = c("hospitalisation", "discharge"),
  # Set a min_delay of 3 days for onset -> hospitalisation
  min_delay = c(3, 7),
  max_delay = c(10, 20)
)

result <- initialise_row(individual_data, delay_map, delay_boundaries)

# hospitalisation identified as incompatible, incompatible date and missing
# date both imputed
result
# id group      onset hospitalisation  discharge
#  1     3 2025-05-01      2025-05-07 2025-05-20

}
