% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_initial}
\alias{mcmc_initial}
\title{Create initial parameter values}
\usage{
mcmc_initial(
  model,
  initial_delay_mean = 7,
  initial_delay_cv = 0.2,
  initial_prob_error = 1
)
}
\arguments{
\item{model}{Model}

\item{initial_delay_mean}{The initial value for the delay means}

\item{initial_delay_cv}{The initial value for the delay coefficients of
variation}

\item{initial_prob_error}{The initial value for the probability of error}
}
\value{
Vector of initial parameter values
}
\description{
Create initial parameter values
}
