% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_run}
\alias{mcmc_run}
\title{Run MCMC}
\usage{
mcmc_run(
  observed_data,
  delays,
  control = mcmc_control(),
  hyperparameters = mcmc_hyperparameters()
)
}
\arguments{
\item{observed_data}{Observed data}

\item{delays}{Delays information}

\item{control}{List of control parameters}

\item{hyperparameters}{List of hyperparameters}
}
\value{
Output
}
\description{
Run MCMC
}
