% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_observation_errors}
\alias{simulate_observation_errors}
\title{Add observation errors}
\usage{
simulate_observation_errors(true_data, error_params, date_range)
}
\arguments{
\item{true_data}{Dataframe containing the true, unobserved dates.}

\item{error_params}{A list containing \code{prop_missing_data} and \code{prob_error}.}

\item{date_range}{A vector of two integer dates for the simulation range.}
}
\description{
Simulate observed data incorporating observation error from true
data
}
