% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_control}
\alias{mcmc_control}
\title{Create control parameters}
\usage{
mcmc_control(
  n_steps = 1000,
  burnin = 0,
  thinning_factor = 1,
  n_chains = 1,
  parallel = FALSE,
  n_workers = 1,
  lower_quantile = 0.01,
  upper_quantile = 0.99
)
}
\arguments{
\item{n_steps}{}

\item{burnin}{}

\item{thinning_factor}{}

\item{n_chains}{}

\item{parallel}{}

\item{n_workers}{}

\item{lower_quantile}{}

\item{upper_quantile}{}
}
\value{
List of control parameters
}
\description{
Create control parameters
}
