% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_control}
\alias{mcmc_control}
\title{Create control parameters}
\usage{
mcmc_control(
  n_steps = 1000,
  burnin = 0,
  thinning_factor = 1,
  n_chains = 1,
  parallel = FALSE,
  n_workers = 1,
  lower_quantile = 0.01,
  upper_quantile = 0.99,
  mean_sdlog = 1,
  cv_sdlog = 1,
  prob_update_estimated_dates = 0.1,
  prob_update_error_indicators = 0.1,
  prob_error_swap = 1
)
}
\arguments{
\item{n_steps}{The number of steps to run in each MCMC chain}

\item{burnin}{The number of steps at the beginning of each chain to discard
as burnin}

\item{thinning_factor}{A thinning factor applied to the chains. If given,
every\code{thinning_factor}'th step is retained}

\item{n_chains}{The number of chains to run}

\item{parallel}{Logical, indicating whether or not to run chains in parallel}

\item{n_workers}{Number of workers to use for parallelisation}

\item{lower_quantile}{Lower quantile used for initialisation of true dates}

\item{upper_quantile}{Upper quantile used for initialisation of true dates}

\item{mean_sdlog}{The sdlog proposal parameter for the delay means}

\item{cv_sdlog}{The sdlog proposal parameter for the delay coefficients of
variation}

\item{prob_update_estimated_dates}{The probability of proposing an update
to each estimated date at each iteration in the MCMC}

\item{prob_update_error_indicators}{The probability of proposing an update
to each error indicator (with the corresponding estimated date updated
accordingly) at each iteration in the MCMC}

\item{prob_error_swap}{The probability of proposing to swap all errors to
non-errors and vice versa (excluding missing dates) for individuals with at
least one error and non-error at each iteration in the MCMC}
}
\value{
List of control parameters
}
\description{
Create control parameters
}
