% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_initial}
\alias{mcmc_initial}
\title{Create initial parameter values}
\usage{
mcmc_initial(
  model,
  initial_mean_delay = 7,
  initial_cv_delay = 0.2,
  initial_prob_error = 1
)
}
\arguments{
\item{model}{Model}

\item{initial_mean_delay}{The initial value for the mean delays}

\item{initial_cv_delay}{The initial value for the coefficient of variation
of delays}

\item{initial_prob_error}{The initial value for the probability of error}
}
\value{
Vector of initial parameter values
}
\description{
Create initial parameter values
}
