% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_run}
\alias{mcmc_run}
\title{Run MCMC}
\usage{
mcmc_run(model, sampler, initial = NULL, control = mcmc_control())
}
\arguments{
\item{model}{Model}

\item{sampler}{Sampler}

\item{initial}{Initial values}

\item{control}{List of control parameters}
}
\value{
Output
}
\description{
Run MCMC
}
