% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a dataset with true dates, observed dates and error indicators}
\usage{
simulate_data(
  n_per_group,
  delay_map,
  delay_params,
  error_params,
  date_range,
  simul_error = FALSE
)
}
\arguments{
\item{n_per_group}{Vector of number of individuals to simulate in each group.}

\item{delay_map}{A data frame that defines the delays between events. It must
contain the columns \code{from} (character), \code{to} (character), and \code{group} (list
of numeric group IDs).}

\item{delay_params}{A data frame containing the parameters (\code{delay_mean},
\code{delay_cv}) for each delay. Consider combining delay_map and delay_params?}

\item{error_params}{A list containing \code{prop_missing_data} and \code{prob_error}.}

\item{date_range}{A vector of two integer dates for the simulation range.}

\item{simul_error}{Boolean. If TRUE, simulates missing and erroneous data.}
}
\value{
A list with three data frames: \code{true_data}, \code{observed_data}, and
\code{error_indicators}, each including \code{id} and \code{group} columns.
}
\description{
Simulate a dataset with true dates, observed dates and error indicators
}
\examples{
# Define the delay_map data frame
delay_map <- data.frame(
  from = c("onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  group = I(list(1:4, 2, 3, 3, 4, 4))
)

# Define the delay parameters data frame
delay_params <- data.frame(
  group = c(1:4, 2, 3, 3, 4, 4),
  from = c("onset", "onset", "onset", "onset", "onset", "onset",
           "hospitalisation", "onset", "hospitalisation"),
  to = c("report", "report", "report", "report", "death", "hospitalisation",
         "discharge", "hospitalisation", "death"),
  delay_mean = c(10, 10, 10, 10, 15, 7, 20, 7, 12),
  delay_cv = c(0.3, 0.3, 0.3, 0.3, 0.4, 0.2, 0.5, 0.2, 0.3)
)

# Define other parameters
n_per_group <- rep(10, max(delay_params$group))
error_params <- list(prop_missing_data = 0.2, prob_error = 0.05)
date_range <- as.integer(as.Date(c("2025-03-01", "2025-09-01")))

# Run simulation
set.seed(1)
sim_result <- simulate_data(
  n_per_group = n_per_group,
  delay_map = delay_map,
  delay_params = delay_params,
  error_params = error_params,
  date_range = date_range,
  simul_error = TRUE
)

sim_result$true_data
sim_result$observed_data
sim_result$error_indicators # true error indicators

}
