% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_true_data}
\alias{simulate_true_data}
\title{Simulate true data}
\usage{
simulate_true_data(n_per_group, delay_map, delay_params, date_range)
}
\arguments{
\item{n_per_group}{Vector of number of individuals to simulate in each group.}

\item{delay_map}{A data frame that defines the delays between events. It must
contain the columns \code{from} (character), \code{to} (character), and \code{group} (list
of numeric group IDs).}

\item{delay_params}{A data frame containing the parameters (\code{mean_delay},
\code{cv_delay}) for each delay. Consider combining delay_map and delay_params?}

\item{date_range}{A vector of two integer dates for the simulation range.}
}
\description{
Simulate the true, unobserved dates for individuals
}
